adjusted datatype from WORD to int for better handling of background subtraction results and to prevent underflow issues. This change was applied across multiple files where spectral data is processed, ensuring that the signal minus background calculations can yield negative values without wrapping around, which is crucial for accurate data analysis in spectroscopic applications.

+++ .src/ConsoleApp.cpp
@@ -369,8 +369,8 @@ int main() {
                 stage->waitOnTarget("1");
                 stage->waitOnTarget("2");
 
-                std::vector<std::vector<std::vector<WORD>>> cube(
-                    nY, std::vector<std::vector<WORD>>(nX, std::vector<WORD>(nPix, 0)));
+                std::vector<std::vector<std::vector<int>>> cube(
+                    nY, std::vector<std::vector<int>>(nX, std::vector<int>(nPix, 0)));
 
                 for (int iy = 0; iy < nY; iy++) {
                     double yPos = yStart + iy * yStep;
@@ -384,7 +384,7 @@ int main() {
                     stage->moveAbs("1", xStop);
 
                     cam->waitForAcquisition();
-                    std::vector<WORD> lineData = cam->getAllSpectra(nX, nPix);
+                    std::vector<int> lineData = cam->getAllSpectra(nX, nPix);
                     for (int ix = 0; ix < nX; ix++) {
                         cube[iy][ix].assign(lineData.begin() + ix * nPix,
                                             lineData.begin() + (ix + 1) * nPix);
